Fix prophet bootstrap
